refactor: replace inline navigation with GameNavBar component for better structure
