+++ src/components/RelativeGameScreen.vue
@@ -1,13 +1,11 @@
 <script setup>
 import { computed, onMounted } from "vue";
-import { useRouter } from "vue-router";
 import { useGlobalGameState } from "../gameState";
 import BaseButton from "./base/BaseButton.vue";
 import BaseSlider from "./base/BaseSlider.vue";
+import GameNavBar from "./GameNavBar.vue";
 
-const router = useRouter();
 const state = useGlobalGameState();
-const mode = state.mode;
 
 // Ensure the relative mode is initialized when component loads
 onMounted(() => {
@@ -30,28 +28,13 @@ const valueDifference = computed({
 function submit() {
 	state.checkGuess(); // Changed from checkRelativeGuess to use the unified API
 }
-
-function goToHome() {
-	router.push("/");
-}
-
-function startOver() {
-	state.startOver(); // Reset the game state
-}
 </script>
 
 <template>
   <div class="flex size-full flex-col justify-between p-2 pb-4">
     <div class="flex flex-col gap-3">
       <!-- Navigation bar -->
-      <div class="flex w-full justify-between">
-        <BaseButton variant="secondary" size="sm" @click="goToHome">
-          <span class="text-lg mr-1">←</span> {{ $t('home') }}
-        </BaseButton>
-        <BaseButton variant="secondary" size="sm" @click="startOver">
-          {{ $t('startOver') }}
-        </BaseButton>
-      </div>
+      <GameNavBar />
 
       <!-- Regular game toolbar with all stats -->
       <Toolbar />
